fix(cli): баннер обновления показывается даже при ошибках команд
Перенос баннера и background refresh из postrun в init hook
через process.on('exit') — postrun не срабатывает при ошибках.
