+++ packages/cli/src/hooks/init.ts
@@ -1,4 +1,10 @@
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { fileURLToPath } from 'node:url';
+import { spawn } from 'node:child_process';
 import { Hook } from '@oclif/core';
+import semver from 'semver';
+import ansis from 'ansis';
 import {
   getActiveProfile,
   getProfile,
@@ -10,7 +16,7 @@ import {
 } from '../profiles.js';
 import { request } from '../client.js';
 import { outputError } from '../output.js';
-import { defaultOutputFormat } from '../utils.js';
+import { defaultOutputFormat, isInteractive } from '../utils.js';
 
 /**
  * Resolve output format from argv before flags are parsed by oclif.
@@ -52,6 +58,56 @@ const SKIP_AUTH = new Set([
 const hook: Hook<'init'> = async function (opts) {
   const commandId = opts.id;
   if (!commandId) return;
+
+  // Register update banner on process exit (fires even when command fails)
+  if (isInteractive() && !process.env.PACHCA_SKIP_NEW_VERSION_CHECK) {
+    const cacheDir = opts.config.cacheDir;
+    const versionFile = path.join(cacheDir, 'version');
+    const currentVersion = opts.config.version;
+
+    process.on('exit', () => {
+      try {
+        if (fs.existsSync(versionFile)) {
+          const cached = JSON.parse(fs.readFileSync(versionFile, 'utf-8'));
+          const latest = cached.latest;
+          if (latest && semver.valid(latest) && semver.valid(currentVersion) && semver.gt(latest, currentVersion)) {
+            const box = [
+              '',
+              ansis.yellow(`╭${'─'.repeat(37)}╮`),
+              ansis.yellow(`│  Доступна новая версия: ${latest.padEnd(12)}│`),
+              ansis.yellow(`│  pachca upgrade${' '.repeat(21)}│`),
+              ansis.yellow(`╰${'─'.repeat(37)}╯`),
+              '',
+            ].join('\n');
+            process.stderr.write(box);
+          }
+        }
+      } catch {
+        // ignore
+      }
+    });
+
+    // Background refresh if cache is stale (>1h)
+    try {
+      let shouldRefresh = true;
+      if (fs.existsSync(versionFile)) {
+        const stat = fs.statSync(versionFile);
+        shouldRefresh = Date.now() - stat.mtimeMs > 60 * 60 * 1000;
+      }
+      if (shouldRefresh) {
+        if (!fs.existsSync(cacheDir)) {
+          fs.mkdirSync(cacheDir, { recursive: true });
+        }
+        const scriptPath = path.join(path.dirname(fileURLToPath(import.meta.url)), '..', 'get-version.js');
+        if (fs.existsSync(scriptPath)) {
+          const child = spawn(process.execPath, [scriptPath, versionFile], { detached: true, stdio: 'ignore' });
+          child.unref();
+        }
+      }
+    } catch {
+      // ignore
+    }
+  }
 
   // Skip auth for exempt commands
   if (SKIP_AUTH.has(commandId)) return;
